Rewrite matcher
The idea is that it'll keep a list of all possible branches, exploring
each one. Then it keeps feeding them, until either all branches end up
dead ends (counts as a failure), or one of the branches got explored all
the way until the end (counts as a success).

Thanks to @tbodt for hinting me how to do this! It's hopefully going to
make implementing groups easier too :)

This may or may not be called NFA or DFA or something cool like that.